==> 2. Cleaning.ipynb <==



<h1><center>In Vino Veritas</center></h1>
<h2><center> <span style="font-weight:normal"><font color='#e42618'> Finding truth in wine data</font>  </span></center></h2>


<h3><center><font color='gray'>JONAS GOTTAL</font></center></h3>




<h4>Project scope</h4>

In 'Vino Veritas' is an old Latin phrase that means 'in wine, there is truth'. And we would like to obtain said truth in wine. What is objectively good wine and how is it influenced by the weather? The Wine Spectators official vintage charts are used as an objective measure for good wine, which is enriched by wine-specific features from public weather data.
<br>

---
---




# Data Cleaning
We need to adjust some of the data, since it isn't coherent. The information is not structured identical for each sub-page, so we make sure the country, region and grape variety is consistent.

# load the data from the JSON file
# data.json
import json
# supress the warnings
import warnings
warnings.filterwarnings("ignore")


with open('data/data.json') as f:
    data = json.load(f)

first we turn it in a dataframe

# turn JSON format into a long table
# with columns: Country, Region, Vintage, Score, Drink Window, Description
import pandas as pd
import numpy as np


# Initialize an empty list to store the DataFrames
dfs = []

# Iterate over the dictionary
for country, regions in data.items():
    for region, tables in regions.items():
        # Convert the nested list into a DataFrame
        df = pd.DataFrame(tables[1:], columns=tables[0])
        # Add the country and region as columns
        df['Country'] = country
        df['Region'] = region
        # Append the DataFrame to the list
        dfs.append(df)

# Concatenate all the DataFrames into a single DataFrame
df = pd.concat(dfs, ignore_index=True)

#drop all rows with non numeric Vintages
df = df[pd.to_numeric(df['Vintage'], errors='coerce').notnull()]

# make a copy of the column Region and call it Variety
df['Variety'] = df['Region']

# Print the DataFrame
df


then we melt the additional columns with either regions or varieties into a long table format.

id_vars = [col for col in df.columns if col not in ['Carneros', 'Napa', 'Santa Barbara', 'Sonoma', "Paso Robles"]]


# Assuming 'df' is your DataFrame and 'Carneros', 'Napa', 'Santa Barbara', 'Sonoma' are your columns
melted_df = df.melt(id_vars=id_vars, 
                    value_vars=['Carneros', 'Napa', 'Santa Barbara', 'Sonoma', "Paso Robles"], 
                    var_name='new_Region', 
                    value_name='new_Score')

# List of columns that you want to keep
id_vars = [col for col in melted_df.columns if col not in ['Cabernet Franc', 'Chenin Blanc', 'Sauvignon Blanc']]

# Use melt function to reshape the DataFrame
melted_df = melted_df.melt(id_vars=id_vars, 
                           value_vars=['Cabernet Franc', 'Chenin Blanc', 'Sauvignon Blanc'], 
                           var_name='new_Variety', 
                           value_name='new_Score_2')


# Print the melted DataFrame
melted_df

we use masks to move the date to the dedicated culomns:


# Create a mask where 'Score' is NaN and 'new_Score' is not NaN
mask = melted_df['Score'].isna() & melted_df['new_Score'].notna()

# Replace NaN values in 'Score' with values from 'new_Score' where the mask is True
melted_df.loc[mask, 'Score'] = melted_df.loc[mask, 'new_Score']

# Replace 'Region' with 'new_Region' where the mask is True
melted_df.loc[mask, 'Region'] = melted_df.loc[mask, 'new_Region']


# Create a mask where 'Score' is NaN and 'new_Score' is not NaN
mask = melted_df['Score'].isna() & melted_df['new_Score_2'].notna()

# Replace NaN values in 'Score' with values from 'new_Score' where the mask is True
melted_df.loc[mask, 'Score'] = melted_df.loc[mask, 'new_Score_2']

# Replace 'Region' with 'new_Region' where the mask is True
melted_df.loc[mask, 'Variety'] = melted_df.loc[mask, 'new_Variety']

and then delete the old columns

# drop where Score is NaN
melted_df = melted_df.dropna(subset=['Score'])
# Remove duplicate rows based on 'Vintage', 'Score', 'Drink Window', 'Description', 'Country', 'Region'
melted_df = melted_df.drop_duplicates(subset=['Vintage', 'Score', 'Country', 'Region'])
# Remove 'new_Region' and 'new_Score' columns
melted_df = melted_df.drop(columns=['new_Region', 'new_Score', 'new_Variety', 'new_Score_2'])
melted_df

# show unique regions
melted_df['Region'].unique()



# for region California/Cabernet/Napa replace with Napa and Variety with Cabernet
melted_df.loc[(melted_df['Region'] == 'California/Cabernet/Napa'), 'Variety'] = 'Cabernet'
melted_df.loc[(melted_df['Region'] == 'California/Cabernet/Napa'), 'Region'] = 'Napa'
# same for Riesling, Germany -> Rheingau (Riesling only Variety)
melted_df.loc[(melted_df['Region'] == 'Riesling'), 'Variety'] = 'Riesling'
melted_df.loc[(melted_df['Region'] == 'Riesling'), 'Region'] = 'Rheingau'
# same for Pinot noir
melted_df.loc[(melted_df['Region'] == 'California/Pinot Noir'), 'Variety'] = 'Pinot Noir'
melted_df.loc[(melted_df['Region'] == 'California/Pinot Noir'), 'Region'] = 'California'
# the same for Oregon: Pinot Noir
melted_df.loc[(melted_df['Region'] == 'Oregon: Pinot Noir'), 'Variety'] = 'Pinot Noir'
melted_df.loc[(melted_df['Region'] == 'Oregon: Pinot Noir'), 'Region'] = 'Oregon'
# the same for 'Washington: Merlot, Cabernet and Syrah'
melted_df.loc[(melted_df['Region'] == 'Washington: Merlot, Cabernet and Syrah'), 'Variety'] = 'Merlot, Cabernet and Syrah'
melted_df.loc[(melted_df['Region'] == 'Washington: Merlot, Cabernet and Syrah'), 'Region'] = 'Washington'
# same for 'Barossa and McLaren Vale: Shiraz'
melted_df.loc[(melted_df['Region'] == 'Barossa and McLaren Vale: Shiraz'), 'Variety'] = 'Shiraz'
melted_df.loc[(melted_df['Region'] == 'Barossa and McLaren Vale: Shiraz'), 'Region'] = 'Barossa and McLaren Vale'
#same for 'Victoria: Shiraz'
melted_df.loc[(melted_df['Region'] == 'Victoria: Shiraz'), 'Variety'] = 'Shiraz'
melted_df.loc[(melted_df['Region'] == 'Victoria: Shiraz'), 'Region'] = 'Victoria'
#same for 'Whites', 'Reds',
melted_df.loc[(melted_df['Region'] == 'Whites'), 'Variety'] = 'Whites'
melted_df.loc[(melted_df['Region'] == 'Reds'), 'Variety'] = 'Reds'
#replace the region of 'Bordeaux/Left Bank Reds (Médoc, Pessac-Léognan)',
melted_df.loc[(melted_df['Region'] == 'Bordeaux/Left Bank Reds (Médoc, Pessac-Léognan)'), 'Region'] = 'Médoc'
#same 'Bordeaux/Right Bank Reds (Pomerol, St.-Emilion)',
melted_df.loc[(melted_df['Region'] == 'Bordeaux/Right Bank Reds (Pomerol, St.-Emilion)'), 'Region'] = 'Saint-Émilion'
#same 'Bordeaux/Sauternes', 
melted_df.loc[(melted_df['Region'] == 'Bordeaux/Sauternes'), 'Region'] = 'Sauternes'
#same 'Bordeaux/Vintage Reds Pre-1995',
melted_df.loc[(melted_df['Region'] == 'Bordeaux/Vintage Reds Pre-1995'), 'Region'] = 'Bordeaux'
#replace Burgundy/Older Vintage Reds with Burgundy
melted_df.loc[(melted_df['Region'] == 'Burgundy/Older Vintage Reds'), 'Region'] = 'Burgundy'
# same 'Burgundy/White',
melted_df.loc[(melted_df['Region'] == 'Burgundy/White'), 'Region'] = 'Burgundy'
#remove "Burgundy/" from all Regions it appears in
melted_df['Region'] = melted_df['Region'].str.replace('Burgundy/', '')
#replace 'Tuscany/Chianti and Chianti Classico with Tuscany, Chianti
melted_df.loc[(melted_df['Region'] == 'Tuscany/Chianti and Chianti Classico'), 'Region'] = 'Tuscany, Chianti'
#same with Tuscany/Bolgheri, Maremma'
melted_df.loc[(melted_df['Region'] == 'Tuscany/Bolgheri, Maremma'), 'Region'] = 'Tuscany, Bolgheri'
# same with 'Tuscany/Brunello di Montalcino'
melted_df.loc[(melted_df['Region'] == 'Tuscany/Brunello di Montalcino'), 'Region'] = 'Brunello di Montalcino'
#replace Tuscany/ with Tuscany,
melted_df['Region'] = melted_df['Region'].str.replace('Tuscany/', 'Tuscany, ')
# for region 'Pinot Noir' put the variety as Pinot Noir
melted_df.loc[(melted_df['Region'] == 'Pinot Noir'), 'Variety'] = 'Pinot Noir'
#same for 'Vintage Port'
melted_df.loc[(melted_df['Region'] == 'Vintage Port'), 'Variety'] = 'Port'
# same for 'Rhône/Northern',and 'Rhône/Southern',
melted_df.loc[(melted_df['Region'] == 'Rhône/Northern'), 'Region'] = 'Rhône'
melted_df.loc[(melted_df['Region'] == 'Rhône/Southern'), 'Region'] = 'Rhône'
#same for 'Côtes de Beaune Reds',
melted_df.loc[(melted_df['Region'] == 'Côtes de Beaune Reds'), 'Variety'] = 'Reds'
melted_df.loc[(melted_df['Region'] == 'Côtes de Beaune Reds'), 'Region'] = 'Côtes de Beaune'
#same for 'Côtes de Nuits Reds',
melted_df.loc[(melted_df['Region'] == 'Côtes de Nuits Reds'), 'Variety'] = 'Reds'
melted_df.loc[(melted_df['Region'] == 'Côtes de Nuits Reds'), 'Region'] = 'Côtes de Nuits'
#same for 'Douro Valley Reds'
melted_df.loc[(melted_df['Region'] == 'Douro Valley Reds'), 'Variety'] = 'Reds'
melted_df.loc[(melted_df['Region'] == 'Douro Valley Reds'), 'Region'] = 'Douro Valley'
# for all regions with 'Whites', 'Reds', 'Vintage Port', 'null', Pinot Noir replace with country
melted_df.loc[(melted_df['Region'] == 'Whites'), 'Region'] = melted_df['Country']
melted_df.loc[(melted_df['Region'] == 'Reds'), 'Region'] = melted_df['Country']
melted_df.loc[(melted_df['Region'] == 'Vintage Port'), 'Region'] = melted_df['Country']
melted_df.loc[(melted_df['Region'] == 'null'), 'Region'] = melted_df['Country']
melted_df.loc[(melted_df['Region'] == 'Pinot Noir'), 'Region'] = melted_df['Country']
# Carneros is Carneros valley
melted_df.loc[(melted_df['Region'] == 'Carneros'), 'Region'] = 'Carneros Valley'

melted_df['Region'].unique()


# save as CSV file
melted_df.to_csv('data/wine_data.csv', index=False)
