==> 5. Modelling.ipynb <==

<h1><center>In Vino Veritas</center></h1>
<h2><center> <span style="font-weight:normal"><font color='#e42618'> Finding truth in wine data</font>  </span></center></h2>


<h3><center><font color='gray'>JONAS GOTTAL</font></center></h3>




<h4>Project scope</h4>

In 'Vino Veritas' is an old Latin phrase that means 'in wine, there is truth'. And we would like to obtain said truth in wine. What is objectively good wine and how is it influenced by the weather? The Wine Spectators official vintage charts are used as an objective measure for good wine, which is enriched by wine-specific features from public weather data.
<br>

---
---


#!conda env export > environment.yml
#!pip freeze | grep -v "^-e" | grep -v "@" | awk -F= '{print $1 "==" $3}' > requirements.txt


# Predictive modelling
We follow our earlier cleaning approach for the merged data:

# import pandas etc
import pandas as pd
import numpy as np

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, r2_score, mean_absolute_error

from xgboost import XGBRegressor
# suppress warnings
import warnings
warnings.filterwarnings('ignore')

# load csv
df = pd.read_csv('data/merged_data.csv')

# Replace non-breaking spaces in column names
df.columns = df.columns.str.replace('\xa0', ' ')

df.head()



# make 'Drink or Hold' == 'Drink or hold'
df['Drink Window'] = df['Drink Window'].str.replace('Drink or hold', 'Drink or Hold')

# interpret as 2, 1, 0, -1, and Nan for  'Drink', 'Hold', 'Drink or Hold', 'Past peak', 'NYR'
df['Drink Window'] = df['Drink Window'].replace('Drink', '2')
df['Drink Window'] = df['Drink Window'].replace('Hold', '1')
df['Drink Window'] = df['Drink Window'].replace('Drink or Hold', '0')
df['Drink Window'] = df['Drink Window'].replace('Past peak', '-1')
df['Drink Window'] = df['Drink Window'].replace('NYR', np.nan) # not yet rated

# also remove the star *:
# Remove trailing asterisks from the "Score" column
df['Score'] = df['Score'].str.rstrip('*')

# Replace 'NYR' with NaN in the "Score" column
df['Score'] = df['Score'].replace('NYR', np.nan)


# Drop rows where 'Score' is NaN or infinite
df = df.dropna(subset=['Score'])
df = df[~df['Score'].isin([np.inf, -np.inf])]

# Define a function to convert a score to a number
def convert_score(score):
    if isinstance(score, str) and '–' in score:
        low, high = map(int, score.split('–'))
        return (low + high) / 2
    else:
        return score
   

# Apply the function to the "Score" column
df['Score'] = df['Score'].apply(convert_score)

# make Vintage, Score, Drink Window a int
df['Vintage'] = df['Vintage'].astype(int)
df['Score'] = df['Score'].astype(float)
df['Drink Window'] = df['Drink Window'].astype(float)


## Basic model: XGBoost

# Drop Vintage,	Drink Window,	Description,	Country,	Region,	Variety
df = df.drop(['Vintage', 'Drink Window', 'Description', 'Country', 'Region', 'Variety'], axis=1)
df.head()

# show nan values in y
df['Score'].isna().sum()

# Separate the target variable and input variables
X = df.drop('Score', axis=1)
y = df['Score']

# Split the data into training set and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Create an XGBoost regression model
model = XGBRegressor()

# Train the model
model.fit(X_train, y_train)

# Make predictions on the test set
y_pred = model.predict(X_test)

# Evaluate the model
r2 = r2_score(y_test, y_pred)
mse = mean_squared_error(y_test, y_pred)
mae = mean_absolute_error(y_test, y_pred)
r = X_train.shape[0]
p = 2
adj_r2 = 1-(((1-r2)*(r-1))/(r-p-1))
print('R-square score: {:.6f}'.format(r2))
print('Adjusted R-square score on test set: {:.6f}'.format(adj_r2))
print("The mean squared error (MSE) on test set: {:.6f}".format(mse))
print("The absolute error (MAE) on test set: {:.6f}".format(mae))

This is just a quick and preliminary modelling approach, but the results are subpar at best. We could already see from the initial analysis that there are no strong predictive signals in the weather data.
